# openapi.yaml
openapi: 3.0.3
info:
  title: RefOnRecord – Resume Verifier API
  description: >
    API for managing resume projects, work experiences, and verification requests.
    Verifiers can be emailed to confirm or reject employment history.
  version: "1.0.0"
  contact:
    name: PWP Spring 2026 Group
    email: Sillah.Babar@student.oulu.fi
servers:
  - url: https://refonrecord.onrender.com/api
    description: Production
  - url: http://localhost:5000/api
    description: Local development
security: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
  responses:
    BadRequest:
      description: Invalid request body or validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, invalid, or expired bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated user is not the owner of this resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    VerificationRequest:
      type: object
      properties:
        request_id:
          type: integer
        experience_id:
          type: integer
        verifier_name:
          type: string
        verifier_position:
          type: string
        verifier_email:
          type: string
          format: email
        verification_token:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, verified, rejected, expired]
        verifier_comment:
          type: string
          nullable: true
        requested_at:
          type: string
          format: date-time
        responded_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
        links:
          type: object
          properties:
            self:
              type: string
            experience:
              type: string
            respond:
              type: string
...
paths:
...
  /users/{user_id}/projects/{project_id}/experiences/{experience_id}/verification-requests/:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
      - name: experience_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Verification]
      summary: List all verification requests for an experience (owner only)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of verification requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VerificationRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Verification]
      summary: Create a verification request and email the verifier (owner only)
      description: >
        Sends an email to the verifier with a link to confirm or reject the
        experience. Token expires after 30 days. Sets experience
        `verification_status` to "pending".
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verifier_name, verifier_position, verifier_email]
              additionalProperties: false
              properties:
                verifier_name:
                  type: string
                  minLength: 1
                  maxLength: 200
                verifier_position:
                  type: string
                  minLength: 1
                  maxLength: 200
                verifier_email:
                  type: string
                  format: email
                  maxLength: 255
      responses:
        "201":
          description: Verification request created and email sent
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationRequest"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{user_id}/projects/{project_id}/experiences/{experience_id}/verification-requests/{request_id}/:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: integer
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
      - name: experience_id
        in: path
        required: true
        schema:
          type: integer
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Verification]
      summary: Get a single verification request (owner only)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Verification request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /verification-requests/{request_id}/respond/:
    parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: integer
    post:
      tags: [Verification]
      summary: Submit a verifier response (public — no login required)
      description: >
        The verifier submits the `verification_token` (from their email link)
        along with "verified" or "rejected". Updates the associated experience
        `verification_status` accordingly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verification_token, status]
              additionalProperties: false
              properties:
                verification_token:
                  type: string
                  format: uuid
                status:
                  type: string
                  enum: [verified, rejected]
                verifier_comment:
                  type: string
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationRequest"
        "400":
          description: Invalid token, expired token, or already responded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Share Links ──────────────────────────────────────────────────────────────
...